% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.mig.r
\name{buffer.mig}
\alias{buffer.mig}
\title{Buffer migration}
\usage{
buffer.mig(land, target.cell.ids)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{target.cell.ids}{A vector of \code{cell.id} codes for those cells open for colonization, 
and for which the presence of source populations nearby must be assessed}
}
\value{
A data frame with the \code{cell.id} of the target cells and the presence (TRUE) or 
absence (FALSE) of a sufficient number of source populations for each 
species: PET, BOJ, ERS, SAB, EPN, other, NonFor
}
\description{
Determines the presence of species within a buffer around target cells.
}
\examples{
data(landscape)
buffer.mig(landscape, landscape$cell.id[runif(10,1,nrow(landscape))])

}
