% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.transition.r
\name{forest.transition}
\alias{forest.transition}
\title{Forest transitions}
\usage{
forest.transition(land, target.cells, suitab, params, type.trans = "S")
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{suitab}{A data frame with climatic suitability and soil suitability for potential colonizing species 
for each forest stand. It is the result of the function \code{suitability()}}

\item{params}{A list of default parameters generated by the function \code{default.params()} or 
a customized list of model parameters}

\item{type.trans}{A character indicating the type of transition: \code{B} for burnt cells, 
\code{O} for cells affected by an outbreak, \code{B} for clear-cut cells, and 
\code{S} for cells undergoing natural succession.}

\item{target.cell.ids}{A vector of \code{cell.id} codes for those cells that may change stand composition}
}
\value{
A vector with the name of the new species / group of species
}
\description{
Determines the presence of species within a buffer around target cells.
}
\examples{
data(landscape)
params <- default.params()
suitab <- suitability(landscape, params) 
forest.trans(landscape, landscape$cell.id[runif(10,1,nrow(landscape))], suitab, params, type.trans = "S")

}
