% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel.type.r
\name{fuel.type}
\alias{fuel.type}
\title{Fuel type}
\usage{
fuel.type(land)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}
}
\value{
A data frame with fuel type and flammability of forest stands
}
\description{
Assign a fuel type and flammability to foret stands according to species and age
}
\examples{

data(landscape)
fuel.type(landscape[runif(10,1,nrow(landscape)),])

}
