% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.volume.r
\name{harvest.volume}
\alias{harvest.volume}
\title{Volume-based clear and partial cuts}
\usage{
harvest.volume(land, params, cc.vol, pc.vol, km2.pixel)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{params}{A list of default parameters generated by the function \code{default.params()} or 
a customized list of model parameters}

\item{cc.vol}{A data frame with the number of cells to clear cut per mgmt unit returned by the function \code{timber.volume()}}

\item{pc.vol}{A data frame with the number of cells to partial cut per mgmt unit returned by the function \code{timber.partial.volume()}}

\item{km2.pixel}{Area in km^2^ of grid cells}
}
\value{
A list of four items: 
 \itemize{
 \item{\code{cc.cells}: A vector with the \code{cell.id} of the clear cut cells}
 \item{\code{pc.cells}: A vector with the \code{cell.id} of the partial cut cells}
 \item{\code{track.cut}: A data frame with managed areas by different prescriptions}
 \item{\code{spp.track}: A data frame with the area and volume clear-cut and partial cut per species}
 }
}
\description{
Selects cells for clear and partial cuts on an volume-based
}
\examples{
data(landscape)
params = default.params()
cc.vol = timber.volume(landscape, params) 
pc.vol = timber.partial.volume(landscape, params, pc.step=5) 
harvest = harvest.volume(landscape, params, cc.vol, pc.vol, km2.pixel=4)

}
