% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{landscape}
\alias{landscape}
\title{Forest landscape features of Quebec province}
\format{
A data frame with 147464 rows and 19 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{x}{X coordinate in Lambert Confromal Conic NAD83}
  \item{y}{Y coordinate in Lambert Confromal Conic NAD83}
  \item{temp}{Historical mean temperature, in ºC}
  \item{prec}{Mean annual precipitation, in mm}
  \item{frz}{Fire regime zone}
  \item{eco.type}{Ecological type}
  \item{bioclim.domain}{Bioclimatic domain}
  \item{mgmt.unit}{Management unit}
  \item{spp}{Tree species or group of species}
  \item{age.matu}{Forest stand maturity age}
  \item{soil.type}{Soil type}
  \item{exclus}{Excluded for timber harvesting}
  \item{tsfire}{Time since last fire, in years}
  \item{tssbw}{Time since last spruce budworm outbreak, in years}
  \item{tsccut}{Time since last clear cut, in years}
  \item{tspcut}{Time since last partical cut, in years}
  \item{tscomp}{Timse since last species change composition, in years}
}
}
\source{
\url{https://mffp.gouv.qc.ca/le-ministere/cartes-plans/}
}
\usage{
landscape
}
\description{
Forest stands characteristics of Quebec province in 2020 at 2 x 2 km of spatial resolution
}
\keyword{datasets}
