% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{prec_rcp45}
\alias{prec_rcp45}
\title{Annual precipitation projections under RCP4.5}
\format{
A data frame with 185479 rows and 17 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{prec20}{Annual precipitation projection for 2020-2024, in mm}
  \item{prec25}{Annual precipitation projection for 2025-2029, in mm}
  \item{prec30}{Annual precipitation projection for 2030-2034, in mm}
  \item{prec35}{Annual precipitation projection for 2035-2039, in mm}
  \item{prec40}{Annual precipitation projection for 2040-2044, in mm}
  \item{prec45}{Annual precipitation projection for 2045-2049, in mm}
  \item{prec50}{Annual precipitation projection for 2050-2054, in mm}
  \item{prec55}{Annual precipitation projection for 2055-2059, in mm}
  \item{prec60}{Annual precipitation projection for 2060-2064, in mm}
  \item{prec65}{Annual precipitation projection for 2065-2069, in mm}
  \item{prec70}{Annual precipitation projection for 2070-2074, in mm}
  \item{prec75}{Annual precipitation projection for 2075-2079, in mm}
  \item{prec80}{Annual precipitation projection for 2080-2084, in mm}
  \item{prec85}{Annual precipitation projection for 2085-2089, in mm}
  \item{prec90}{Annual precipitation projection for 2090-2094, in mm}
  \item{prec95}{Annual precipitation projection for 2095-2099, in mm}
}
}
\source{
\url{https://www.ouranos.ca/en/}
}
\usage{
prec_rcp45
}
\description{
Projections of annual precipitation for Quebec province at 2 x 2 km of spatial resolution and 5-year temporal 
resolution under the RCP4.5 scenario derived from the MIROC_ESM_CHEM global model.
}
\keyword{datasets}
