% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLDM.r
\docType{package}
\name{QLDM}
\alias{QLDM}
\title{QLDM: A package for simulating forest landscape dynamics in Quebec province}
\description{
The QLDM package provides a set of functions to simulate landscape-scale processes that 
influence the structure and composition of forest stands in a region, such as forest management 
(timber harvesting, salvagge logging, ...), wildfires and insect outbreaks.
}
\section{Author(s)}{

\bold{Maintainer}: Núria Aquilué \email{nuria.aquilue@ctfc.cat}  

\bold{Authors}: Mathieu Bouchard and Núria Aquilué
}

