% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quebec.ldm.r
\name{quebec.ldm}
\alias{quebec.ldm}
\title{Quebec Landscape Dynamic Model}
\usage{
quebec.ldm(
  is.wildfires = FALSE,
  is.sbw = FALSE,
  is.harvesting = FALSE,
  custom.params = NA,
  rcp = NA,
  prec.proj = NA,
  temp.proj = NA,
  timber.supply = "area.based",
  pigni = NA,
  nrun = 1,
  time.step = 5,
  time.horizon = 80,
  save.land = FALSE,
  out.seq = NA,
  out.path = NA,
  ...
)
}
\arguments{
\item{is.sbw}{A flag to indicate that spruce budworm outbreaks are a simulated process of the model}

\item{is.harvesting}{A flag to indicate that harvesting by clear and partial cuts is a simulated process of the model}

\item{custom.params}{List with the model paramaters and default and/or user-defined values}

\item{rcp}{Climate projection, either \code{NA} (default), 'rcp45' or 'rcp85'}

\item{prec.proj}{Data frame with annual precipitation projections for each time step}

\item{temp.proj}{Data frame with mean temperature projections for each time step}

\item{timber.supply}{Approach to do timber supply calculation, either \code{area.based} or \code{volume.based}}

\item{pigni}{Optional data frame with probability of fire igntion per grid cell}

\item{nrun}{Number of replicates to run the model}

\item{time.step}{Number of years of each time step}

\item{time.horizon}{Number of years of the model simulation, it has to be a multiple \code{time.step}}

\item{save.land}{A flag to save as a RDS file the \code{landscape} data frame at the time step indicated in \code{out.seq}}

\item{out.seq}{Numeric vector with the time steps the \code{landscape} is saved}

\item{out.path}{String with the directory path to save the \code{landscape} data frame at each time step indicated in \code{out.seq}}

\item{is.widlfire}{A flag to indicate that wildfires are a simualted process of the model}
}
\value{
A list with the following items:
 \itemize{
   \item{\code{SppByAgeClass}: A data frame of species abundance per age class, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{mgmt.unit}: Code of the forest management unit (FMU).}
        \item{\code{spp}: Code of the species.}
        \item{\code{age.class}: Code of the age class, \code{C10}, \code{C30}, \code{C50}, \code{C70}, \code{C90}, and \code{OLD}.}
        \item{\code{area}: Area in \eqn{km^{2}}.}
      }
   }
   \item{\code{SuitabilityClass}: A data frame of suitability of potential species per bioclimatic domain, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{bioclim.domain}: Code of the bioclimatic domain.}
        \item{\code{spp}: Code of the species.}
        \item{\code{poor}: Area in \eqn{km^{2}} of poor environmental suitability.}
        \item{\code{med}: Area in \eqn{km^{2}} of intermediate environmental suitability.}
        \item{\code{good}: Area in \eqn{km^{2}} of good environmental suitability.}
      }
   }
   \item{\code{SppByFireZone}: A data frame of species abundance per fire zone, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{spp}: Code of the species.}
        \item{\code{area}: Area in \eqn{km^{2}}.}
      }
   }
   \item{\code{FuelByFireZone}: A data frame of fuel type per fire zone, with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{type}: Code of the fuel type:  \code{low}, \code{med} or \code{high}.}
        \item{\code{pct}: Relative abundance of the fuel type in the fire regime zone ([0,1]).}
      }
   }
   \item{\code{Cuts}: A data frame of harvestable area and volume per management unit 
   (included if \code{is.harvesting}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{mgmt.unit}: Code of the forest management unit (FMU).}
        \item{\code{a.inc}: Area that can be harvested, i.e. non-protected and FMU informed (in \eqn{km^{2}}).}
        \item{\code{a.even.age}: Area within \code{a.inc} of mature stand, i.e.\code{age>age.matu} (in \eqn{km^{2}}).}
        \item{\code{a.mat.pc}: Area to be managed under a partial-cut regime (in \eqn{km^{2}}).}
        \item{\code{a.inc.burnt}: Area within \code{a.inc} burnt in the current time step (in \eqn{km^{2}}).}
        \item{\code{a.inc.mat.burnt}: Area within \code{a.inc} of mature stands burnt in the current time step (in \eqn{km^{2}}).}
        \item{\code{a.inc.kill}: Area within \code{a.inc} killed by SBW in the previous or current time step (in \eqn{km^{2}}).}
        \item{\code{a.inc.mat.kill}: Area within \code{a.inc} of mature stands killed by SBW in the previous or current time step (in \eqn{km^{2}}).}
        \item{\code{a.reg.fail.ex}: Protected area at defforestation risk (in \eqn{km^{2}}).}
        \item{\code{a.reg.fail.in}: Non-protected area at defforestation risk (in \eqn{km^{2}}).}
        \item{\code{a.salvaged}: Salvaged and clear-cut area (in \eqn{km^{2}}).}
        \item{\code{a.unaff}: Clear-cut area unaffected by disturbances (in \eqn{km^{2}}).}
        \item{\code{v.salv}: Salvaged and clear-cut volume (in \eqn{m^{3}}).}
        \item{\code{v.unaff}: Clear-cut volume unaffected by disturbances (in \eqn{m^{3}}).}
        \item{\code{a.pcut}: Area partial cut (in \eqn{km^{2}}).}
        \item{\code{v.pcut}: Volume partial cut (in \eqn{m^{3}}).}
      }
   }
   \item{\code{SppCut}: A data frame of area and volum extracted by clear and partial cut per species 
   and management unit (included if \code{is.harvesting}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{mgmt.unit}: Code of the forest management unit (FMU).}
        \item{\code{spp}: Code of the species.}
        \item{\code{area.ccut}: Clear-cut area (in \eqn{km^{2}}).}
        \item{\code{vol.ccut}: Clear-cut volume (in \eqn{m^{3}}).}
        \item{\code{area.pcut}: Partial cut area (in \eqn{km^{2}}).}
        \item{\code{vol.pcut}: Partial cut volume (in \eqn{m^{3}}).}
      }
   }
   \item{\code{BurntRates}: A data frame of target area to be burnt per fire regime zone 
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{br}: Code of the species.}
        \item{\code{brvar}: Code .}
        \item{\code{brfuel}: Area in \eqn{km^{2}}.}
        \item{\code{brclima}: Area in \eqn{km^{2}}.}
        \item{\code{target.area}: Area in \eqn{km^{2}}.}
      }
   }
   \item{\code{FireRegime}: A data frame of number of fires and burnt area per fire regime zone
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{target.area}: Target area to be burnt in \eqn{km^{2}}.}
        \item{\code{nfires}: N.}
        \item{\code{burnt.area}: Area in \eqn{km^{2}}.}
        \item{\code{fire.cycle}: Area in \eqn{km^{2}}.}
        \item{\code{indx.combust}: Area in \eqn{km^{2}}.}
        \item{\code{indx.combust.burnt}: Area in \eqn{km^{2}}.}
      }
   }
   \item{\code{Fires}: A data frame of wildfires
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{fire.id}: Wildfire identificator.}
        \item{\code{wind}: Main wind direction in degrees.}
        \item{\code{target.size}: Target size in pixels to be burnt.}
        \item{\code{burnt.size}: Burnt size in pixels.}
        \item{\code{rem}: Remanent number of pixels to be burnt.}
      }
   }
   \item{\code{BurntFuels}: A data frame of burnt fuel types per fire regime zone
   (included if \code{is.wildfire}), with columns:
     \itemize{
        \item{\code{run}: Number of replicate.}
        \item{\code{year}: Year YYYY.}
        \item{\code{frz}: Code of the fire regime zone.}
        \item{\code{type}: Code of the fuel type:  \code{low}, \code{med} or \code{high}.}
        \item{\code{area}: Area burnt in \eqn{km^{2}}.}
      }
   }
 }
}
\description{
Run the landscape dynamic model QLDM that includes the processes of wilfires, clear-cuts,
partial-cuts, spruce budworm outbreaks, post-disturbance regeneration, and forest seral succession.
}
\examples{

\dontrun{
library(QLDM)
data(prec)
data(temp)
# Run one single 80-year replicate with forest management and default RCP 4.5 climate projections 
quebec.ldm(is.harvesting = T, rcp = "rcp45")
}

}
