% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbw.outbreak.r
\name{sbw.outbreak}
\alias{sbw.outbreak}
\title{Srpuce budworm outbreak}
\usage{
sbw.outbreak(land, severity = 1)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{severity}{The level of outbreak severity ([0,1])}
}
\value{
A vector with the \code{cell.id} of the locations killed by the outbreak
}
\description{
Simulates spruce budworm mortality according to a given probability of mortality after defoliation
}
\examples{
data(landscape)
kill.cells = sbw.outbreak(landscape)

}
