% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand.volume.r
\name{stand.volume}
\alias{stand.volume}
\title{Stand volume}
\usage{
stand.volume(land)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}
}
\value{
A data frame with the volume per species per management unit
}
\description{
Determine volume as a function of stand age and site index
}
\examples{
data(landscape)
vol = stand.volume(landscape)

}
