% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitability.r
\name{suitability}
\alias{suitability}
\title{Suitability}
\usage{
suitability(land, params)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{params}{A list of default parameters generated by the function \code{default.params()} or 
a customized list of model parameters}
}
\value{
A data frame with enviornmental suitability ([0,1]) of forest stands
}
\description{
Determines the environmental suitability of forest stands according to temperature, precipiation and soil type constraints
}
\examples{
data(landscape)
params = default.params()
fuel.type(landscape[runif(10,1,nrow(landscape)),], params)

}
