% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{temp_rcp45}
\alias{temp_rcp45}
\title{Mean temperature projections under RCP4.5}
\format{
A data frame with 185479 rows and 17 variables:
\describe{
  \item{cell.id}{Unique grid cell indentificator}
  \item{prec20}{Mean temperature projection for 2020-2024, in ºC}
  \item{prec25}{Mean temperature projection for 2025-2029, in ºC}
  \item{prec30}{Mean temperature projection for 2030-2034, in ºC}
  \item{prec35}{Mean temperature projection for 2035-2039, in ºC}
  \item{prec40}{Mean temperature projection for 2040-2044, in ºC}
  \item{prec45}{Mean temperature projection for 2045-2049, in ºC}
  \item{prec50}{Mean temperature projection for 2050-2054, in ºC}
  \item{prec55}{Mean temperature projection for 2055-2059, in ºC}
  \item{prec60}{Mean temperature projection for 2060-2064, in ºC}
  \item{prec65}{Mean temperature projection for 2065-2069, in ºC}
  \item{prec70}{Mean temperature projection for 2070-2074, in ºC}
  \item{prec75}{Mean temperature projection for 2075-2079, in ºC}
  \item{prec80}{Mean temperature projection for 2080-2084, in ºC}
  \item{prec85}{Mean temperature projection for 2085-2089, in ºC}
  \item{prec90}{Mean temperature projection for 2090-2094, in ºC}
  \item{prec95}{Mean temperature projection for 2095-2099, in ºC}
}
}
\source{
\url{https://www.ouranos.ca/en/}
}
\usage{
temp_rcp45
}
\description{
Projections of mean temperature for Quebec province at 2 x 2 km of spatial resolution and 5-year temporal 
resolution under the RCP4.5 scenario derived from the MIROC_ESM_CHEM global model.
}
\keyword{datasets}
