% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timber.area.r
\name{timber.area}
\alias{timber.area}
\title{Sustainable clear-cut area}
\usage{
timber.area(land, params)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{params}{A list of default parameters generated by the function \code{default.params()} or 
a customized list of model parameters}
}
\value{
A data frame with the number of cells to clear cut per mgmt unit with actual names of management units
}
\description{
Calculates sustained-yield levels of clear cuts for each management unit in area
}
\examples{
data(landscape)
params = default.params()
cc.area = timber.area(landscape, params) 

}
