% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timber.partial.volume.r
\name{timber.partial.volume}
\alias{timber.partial.volume}
\title{Sustainable partial cut volume}
\usage{
timber.partial.volume(land, params, pc.step)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{params}{A list of default parameters generated by the function \code{default.params()} or 
a customized list of model parameters}

\item{pc.step}{Number of years of partial cut time step}
}
\value{
A data frame with the number of cells to partial cut per mgmt unit with actual names of management units
}
\description{
Calculates sustained-yield levels of partial cuts for each management unit in volume
}
\examples{
data(landscape)
params = default.params()
pc.vol = timber.partial.volume(landscape, params, pc.step=5) 

}
