% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timber.volume.r
\name{timber.volume}
\alias{timber.volume}
\title{Sustainable clear-cut volume}
\usage{
timber.volume(land, params, time.step = 5)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{params}{A list of default parameters generated by the function \code{default.params()} or 
a customized list of model parameters}

\item{time.step}{Number of years of each time step}
}
\value{
A data frame with the number of cells to clear cut per mgmt unit with actual names of management units
}
\description{
Calculates sustained-yield levels of clear cuts for each management unit in volume
}
\examples{
data(landscape)
params = default.params()
cc.vol = timber.volume(landscape, params) 

}
