% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildfires.r
\name{wildfires}
\alias{wildfires}
\title{Wildfires}
\usage{
wildfires(
  land,
  params,
  baseline.fuel,
  mem.feux,
  rcp = NA,
  pigni = NA,
  km2.pixel = 4,
  time.step = 5,
  t = 5
)
}
\arguments{
\item{land}{A \code{landscape} data frame with forest stand records in rows}

\item{params}{A list of default parameters generated by the function \code{default.params()} or 
a customized list of model parameters}

\item{baseline.fuel}{asdf}

\item{mem.feux}{asdfv}

\item{rcp}{Climate projection, either \code{NA} (default), 'rcp45' or 'rcp85'}

\item{pigni}{Vector with probability of fire ignition ([0,1]) for each cell grid}

\item{km2.pixel}{Area in km^2^ of grid cells}

\item{time.step}{Number of years of each time step}

\item{t}{Current time step}
}
\value{
A list of six items: 
 \itemize{
 \item{\code{brunt.cells}: A vector with the \code{cell.id} of the burnt cells}
 \item{\code{track.target}: A vector with the \code{cell.id} of the partial cut cells}
 \item{\code{track.regime}: A data frame with managed areas by different prescriptions}
 \item{\code{track.fire}: A data frame with the area and volume clear-cut and partial cut per species}
 \item{\code{track.sprd}: A data frame with the area and volume clear-cut and partial cut per species}
 \item{\code{mem.feux}: ¿?}
 }
}
\description{
Simualtes fire events in fire regimes zones
}
\examples{
data(landscape)
params = default.params()
# Compute the baseline fuel at the fire zone level
fuels = fuel.type(land)
baseline.fuel = group_by(fuels, frz) \%>\% summarise(x=mean(flam))
mem.feux = rep(0, length(unique(landscape$frz)))
# Simulate wildfires in all fire regime zones
fires = wildfires(landscape, params, baseline.fuel, mem.feux, rcp = NA,
 pigni = NA, km2.pixel = 4, time.step = 5, t = 5)

}
